Fix data migration to populate mailing references
The data migration that populates references to mailings failed because
the Content Types model was empty. This happens not just when running
unit tests, but merely when a new environment is set up. I encountered
this when setting up a new virtual machine.

This patch adapts the migration logic accordingly.